requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.54.0
python-dotenv==1.0.0
pydantic>=2.10.0
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0
pillow>=10.2.0
pyqt6>=6.10.1
pyinstaller>=6.15.0
aiofiles==23.2.1
python-multipart==0.0.6

